Empty string instead of meaningless object
Removed the meaningless object with no property value. Made no sense to me now. Was in effect with -AllowDuplicates and multiple ID strings. Considering revising this further, but I have a problem thinking of a good, easy-to-parse pattern that also is friendly to the human eye. Likely a compromise will be made in the next revision, for a more pleasant presentation of the data in this situation.